requirements.txt
cohere
pypdf
pdfplumber
chromadb
streamlit
pandas
python-dotenv
